refactor(aipanel): 拆出面板状态基座与设置状态簇 hook
- 新增 components/ai/useAIPanelCoreState.ts：terminalPanels 状态机/setPanelState、composer 草稿、token 账本（全量重建/增量刷新）、会话快照保存、滚动信号与生命周期 ref
- 新增 components/ai/useAIPanelSettingsState.ts：MCP 服务信息与输出压缩设置、全局 AI 设置、设置浮层 UI 态、临时会话/主题预览、提示音与删除确认
- AIPanel.tsx 4692 -> 4108 行
